closed ticket #65.
New file is created "testoob", which is an executable python file, and it runs the suite file given to it using testOOB. Nothing special is needed inside the excuted test file.

Also, this file is now installed when "python setup.py install" is ran, and you can run it just using "testoob" from anywhere.