Closed ticket #16
Added the vassert option.
Now when running tests with --vassert, the asserts are printed to the user.

changes:
verbalize.py - New file.
main.py - Added the option '--vassert' to the command line args.
running.py - Added the call to verbalize, if needed.
reporting.py - Added the option to report and print verbouse asserts.